Create mutable ropes with the right encoding
Previously, replaceExact could inadvertently replace ranges with
incompatible code ranges or encodings.

This commit puts the right checks in place and adds a parameter to
getMutable that allows the right type of node to be created from the start.

This required the runtime assertions in the LeafRope constructors to be
relaxed.