Force splitting of []= and profile each call site
The implementation of `String#[]=` was moved to StringNodes. The default
string_slice `@Specialization` now favors mutability. This is to allow
correct profiling of the case `s[x] = '..'; s[y] = '..'`. The profiling
node now collapses into a constant after an implementation is picked.
The AST size of `[]=` is 13, which might be significant.